# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverrideSet": {
        "properties": {
          "allow": {
            "default": 0,
            "maximum": 16383.0,
            "minimum": 0.0,
            "title": "Allow",
            "type": "integer"
          },
          "deny": {
            "default": 0,
            "maximum": 16383.0,
            "minimum": 0.0,
            "title": "Deny",
            "type": "integer"
          },
          "role_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "title": "OverrideSet",
        "type": "object"
      },
...
      "RoleCreate": {
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "default": 0,
            "maximum": 16383.0,
            "minimum": 0.0,
            "title": "Permissions",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "RoleCreate",
        "type": "object"
      },
      "RoleListOut": {
        "properties": {
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleOut"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "roles"
        ],
        "title": "RoleListOut",
        "type": "object"
      },
      "RoleOut": {
        "properties": {
          "capability_names": {
            "items": {
              "type": "string"
            },
            "title": "Capability Names",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_everyone": {
            "title": "Is Everyone",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "title": "Permissions",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "position",
          "permissions",
          "capability_names",
          "is_everyone"
        ],
        "title": "RoleOut",
        "type": "object"
      },
      "RolePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "permissions": {
            "anyOf": [
              {
                "maximum": 16383.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permissions"
          }
        },
        "title": "RolePatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Openvoice API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/channels/{channel_id}/overrides": {
      "put": {
        "operationId": "set_override_api_v1_channels__channel_id__overrides_put",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Channel Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Set Override Api V1 Channels  Channel Id  Overrides Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Override",
        "tags": [
          "roles"
        ]
      }
    },
...
    "/api/v1/communities/{community_id}/members/{user_id}/roles/{role_id}": {
      "delete": {
        "operationId": "unassign_role_api_v1_communities__community_id__members__user_id__roles__role_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Unassign Role Api V1 Communities  Community Id  Members  User Id  Roles  Role Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unassign Role",
        "tags": [
          "roles"
        ]
      },
      "put": {
        "operationId": "assign_role_api_v1_communities__community_id__members__user_id__roles__role_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Assign Role Api V1 Communities  Community Id  Members  User Id  Roles  Role Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/communities/{community_id}/roles": {
      "get": {
        "operationId": "list_roles_api_v1_communities__community_id__roles_get",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Roles",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "create_role_api_v1_communities__community_id__roles_post",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Role",
        "tags": [
          "roles"
        ]
      }
    },
...
    "/api/v1/roles/{role_id}": {
      "delete": {
        "operationId": "delete_role_api_v1_roles__role_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Role Api V1 Roles  Role Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Role",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "update_role_api_v1_roles__role_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Role",
        "tags": [
          "roles"
        ]
      }
    }
  }
}